// server/src/lightning/types.ts
export interface LightningGap {
  fromMs: number;
  toMs: number;
}